add check for games with same id

+++ game.py
@@ -4,6 +4,7 @@ import deck
 import ability
 import random
 import json
+from exceptions import EvolutionServerException
 
 class Game:
     def __init__(self, name, players, deck):
@@ -20,8 +21,10 @@ class Game:
         self.dice = 0
         self.food = 0
         self.deck = deck
-        self.save()
-        print("Game {} created.".format(self.id))
+        if (self.save()):
+            print("Game {} created.".format(self.id))
+        else:
+            raise EvolutionServerException('Game with this ID already exists!')
 
     def do_evolution(self, player, creature, card):
         if not (self.turn == player.name):
@@ -58,9 +61,12 @@ class Game:
         self.turn = first.name
 
     def save(self):
+        if (os.path.isfile("games/{}.json".format(self.id))):
+            return False
         game = self.json()
         with open('games/{}.json'.format(self.id), 'w') as outfile:
             json.dump(game, outfile)
+            return True
 
     def json(self):
         json = {}
